% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{agingtrees}
\alias{agingtrees}
\title{agingtrees}
\format{
A data frame with 176 rows and 3 variables:
\describe{
  \item{\code{Individual}}{integer Tree ID}
  \item{\code{Age}}{integer Age in years}
  \item{\code{Size}}{double Height in meters }
}
}
\source{
\url{http://webhome.auburn.edu/~tds0009/WILD7150/Assignments/Assign6a.csv}
}
\usage{
agingtrees
}
\description{
Dataset 3: Repeated meausures of tree height of 16 trees over 10 yrs to understand growth trajectory
}
\keyword{datasets}
