% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{hormonaltrees}
\alias{hormonaltrees}
\title{hormonaltrees}
\format{
A data frame with 32 rows and 4 variables:
\describe{
  \item{\code{Field}}{integer Field ID}
  \item{\code{Fertilizer}}{integer Fertilizer added or not added to the field(1/0)}
  \item{\code{Hormone}}{integer Hormone added or not added to the field(1/0)}
  \item{\code{Size}}{double Average tree height in centimeters}
}
}
\source{
\url{http://webhome.auburn.edu/~tds0009/WILD7150/Assignments/Assign6c.csv}
}
\usage{
hormonaltrees
}
\description{
Dataset 2: Average tree height per plot, based on plot exposure to fertilizer and/or hormones
}
\keyword{datasets}
