% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{lotuscountry}
\alias{lotuscountry}
\title{lotuscountry}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{\code{Latitude}}{double Latitude in degrees (continuous)}
  \item{\code{Elevation}}{integer Elevation in meter (continuous) }
  \item{\code{Country}}{character Country lotus data was collected: Nepal, India, Tibet}
  \item{\code{Size}}{double Lotus plant size in grams}
}
}
\source{
\url{http://webhome.auburn.edu/~tds0009/WILD7150/Assignments/Assign6b.csv}
}
\usage{
lotuscountry
}
\description{
Dataset 4: Lotus size (grams) from 3 different countries (w/ continuous variables for elevation and latitude)
}
\keyword{datasets}
