% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{pickygrasshoppers}
\alias{pickygrasshoppers}
\title{pickygrasshoppers}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{\code{Field}}{integer FIELD ID}
  \item{\code{Plot}}{integer Treatments to Plot: Control (nothing), dBurn (Burned), Fertilize (Fertilized) }
  \item{\code{Density}}{double Hopper Density: Number of grasshoppers per hectare}
}
}
\source{
\url{http://webhome.auburn.edu/~tds0009/WILD7150/Assignments/Assign6b.csv}
}
\usage{
pickygrasshoppers
}
\description{
Dataset 1: Grasshopper density (grasshopper/hectare) in response to controlled burn or fertilizer
}
\keyword{datasets}
